결제/취소 브랜치와 주문내역 브랜치 merge후 mainVC에 UI정리
